Use cenipawn integers for eval, compiler optimizations, exception handling in move lookup

+++ Chess/bd.h
@@ -13,7 +13,7 @@
 
 
 
-typedef float EVAL;
+typedef int32_t EVAL;
 
 
 /*
